+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Changed the HTTP return codes for the generic error handlers in Clowder.
 
 ### Added
+- When a folder has been deleted, clowder will traverse each file (directly/indirectly) under this folder and send file deletion event to Rabbitmq.
+  [CATS-995](https://opensource.ncsa.illinois.edu/jira/browse/CATS-995)
 - Script to cleanup/migrate userpass account data to cilogon accounts.
 
 ## 1.6.0 - 2019-04-01